FSPT-781: Allow custom templates for managed expression form rendering
As we build out the context-aware conditions and validation flows, we know that
we'll need some non-trivial form rendering logic (to switch between users
inserting static values and having configured expression statements). To allow
this logic to be isolated in templates, we provide the ability to set a managed
expression form template.