FSPT-781: Allow contextual data in managed expressions
Allows the insertion of data into numerical and date managed expressions
to take the place of arbitrary user-entered values, allowing for
cross-form validation and conditions.

As with the insert data flow for question text/hint/guidance, we save
the form state when the user goes off to the insert contextual data flow
so that when they come back we can restore the form state (ie. the
chosen managed expression and any inclusive checkboxes or other values)

We also conditionally reveal the uneditable qid field when the user
has selected data to insert, to avoid confusion as to which value will
be used in the expression.

The qid is saved with `(())` so that it can be interpolated, and these
are ignored by the expression engine when evaluating the expression
statement.

To do in subsequent commits:
- update add context flow template back/cancel links
- update _validate_and_sync_expression_references to create new rows for
context data component references
- condition select dropdown should only show questions before the
question being depended upon, not the question being edited
- TESTS