FSPT-781: Simplify referencing data in expressions
Now that all 'reference data' buttons are called 'add_context' as form
fields we no longer need to loop over the fields in the
is_submitted_to_add_context check and can instead just reference the
add_context field directly.

`_prepared_form_data` now takes care of rebuilding the form to the state
it was in when you come back from the `reference data` flow, so the
additional form rebuilding in the route was redundant.