FSPT-901: Bootstrap Flask-Admin for platform admin pages
This patch integrates a baseline for Flask-Admin with a simple model view
exposed to list users in the platform. We apply some default
authorisation/authentication checks to validate that only platform admin users
can see this part of the service.

For now this will be based on the ADMIN role in the service, which correlates to
membership of the FSD_ADMIN AD group. We plan to create a FSD_PLATFORM_ADMIN AD
group soon to allow us to further limit who can use platform admin pages, as
FSD_ADMIN is more than just developers/privileged people.