FSPT-883: represent add another form and submission data as lists
This commit preserves fetching answers from the persisted submission
data as a single answer at a time but updates the contexts that data is
then used (by the form runner/ submission pages/ expression evaluation/
data export) to storing a list of all of the answers to that question
for each entry in the add another list.

This should work uniformly for individual questions that are add another
and groups containing questions.

This avoids us flatenning all of the add another answers and starting to
index identifiers which we anticipate will make managing expressions
complicated when stacking with all of the other features of expressions.

Although the add another question ID will be associated with a list
we'll allow the submission helper to add an "Add another context" if it
needs certain expressions to be evaluated for a specific indexed entry
to the add another. This should work consistently with the ChainMap
implementation of expression contexts. It will allow questions within an
add another group to refer to the questions within that same group that
came before them and allow the add another context dict to pick just the
value at the current index for the expression to be evaluated against.