FSPT-894: clean up add another expression context
Tweaks how the expression context is built for add another to remove a
lot of the conditional/ early return behaviour and instead only extend
the context if certain criteria are met.

This PR also adds some lower level/ unit coverage to validate that the
appropriate context is set and expressions referencing within add
another, this will probably replace the higher level integration tests
until all of the functionality is in and we add this back in
intentionally.

Note the way this has been set up is to provide the add another context
when an expression will be evaluated for conditions but if this
behaviour needs to be suppported elsewhere (i.e referencing data in a
validation within that add another group) this logic should probably
move inside the expression context builder itself (accepting an add
another index and adding the context itself) - this would allow for it
to be applied consistently.

There should also be a task at the end of these commits that reviews all
of this behaviour factoring in any group in group specific edge cases we
might run into.