FSPT-781: Refactor session model type checking
We'd added in some route-level session model type checking to avoid
instances of session model overlap (eg. coming back to an expression
form page when adding a validation or condition and then hitting 'back'
to go back to edit question - this used to lead to session overlap
until we added the route-level `if not isinstance` checks to make sure
we had got back the session type we wanted).

This wasn't ideal as it meant there were now multiple places that were
having to deal with session types and clearing out the session or the
add_context_data that was being used - multiple routes and also the
`_extract_add_context_data_from_session` helper method, when ideally
just the latter would deal with handling all of this and the routes
could just expect the session model they should be dealing with or None.

This change shifts that responsibility back to
`_extract_add_context_data_from_session` exclusively and deals with any
session mismatch and clearing out the session when needed.