FSPT-781: Restrict select context source dropdown
When we're selecting a context source for question text/hint or guidance
we generally want all previous questions in the form to be available as
context (with the exception of same-page questions in a group).

However when selecting a context source for expressions we should only
show questions of the same question type (integers for integers, dates
for dates) for the expression evaluation to not be completely broken.
This change allows a boolean to be passed to the form set up to identify
if it's being used for an expression, and if so restricts the question
list to those of the same datatype.

Also a small fix so if we're referencing data for an expression we pass
in the `depends_on_question` as the current component for the dropdown
so form designers can't accidentally select questions that come later
in the form/the depends on question itself.