FSPT-894: update are all questions answered to check lists
If a question is part of an add another question or group the utility
methods should expect to validate that each of the entries that have
been provided for that question are correctly answered (where before we
where just checking for a K:V pair existing).

In order to support questions referring to previous questions in the
same add another group in conditions and validations we make sure the
expression context is appropriate for each entry we're checking.