BAU: Filter condition question select dropdown
Currently the select question dropdown when adding a condition displays
all questions in a form as long as they don't have the same ID as the
current component and are of a question type that can be used for
conditions (eg. integers, dates, yes/no). However it doesn't currently
filter out questions that come after the current component in the form
or questions that are part of the same 'same page' group as the current
component.

Selecting these does trigger validation errors but it could be a little
confusing to the user to see these in the first place, and as we've
started implementing this filtering for the Select Context Source question
dropdown it makes sense to centralise this logic and enforce this
in both places, and then allow additional checks for each where needed.