FSPT-894: update all questions answered interface to not return answers
No consumers of this method are actually processing the answers that are
returned and it makes the code less intuitive to be storing them (and
looks like it sets a precedent even though its not used). Remove it and
return something lightly structured so it can be used where checking the
truthyness of answers was before.