FSPT-894: are all questions answered only calculate context once per entry
Small optimisation to only extend the context once per entry for a given
add another answer, if there are lots of questions in the group or lots
of entries in the add another this should mean doing a lot less work.

It does mean that the "add_another_index" util inside of the is
component visible gets skipped but that could still be useful for things
calling that method that don't already have their predictable context
set up (and will error if you try and set the add another context twice
by accident).