Adding post generation hooks to QuestionFactory and GroupFactory. These ensure that if we call build() (rather than create()), that the form.components property behaves as expected.
When a form object is properly hooked up to the database (when running for real, or in integration tests) the primaryjoin on components ensures that form.components only contains immediate children of the form. But when we call factory.build() on a question or group, there isn't a real database, so the join doesn't work in the same way.

The post generation hooks added in this commit make the factory generated objects filter components in the same was as the join on real database objects, ensuring that unit tests behave the same way as integration tests.