FSPT-907: refactor to avoid namespacing add another lists in contexts
In the future we'll need to make the lists of each add another question
answers available to expressions so they can do aggregate calculations on
them (for example sum all of the reported spending for a quarter), for
this reason we had previously made each answer to a given add another
question available in a list.

This lists result was previously stored against the questions identifier
(similarly to flat simple answers) and would result in a `TypeError` when
expressions are evaluated that hadn't been given a specific "add another
index" to select.

As the list isn't yet consumed we've decided to just remove exposing the
list of all the answers to the context and only fetch add another
answers if an add another context is being requested explicitly - this
does the same thing as separately indexing the list of answers (the
answer is not present unless a context has been requested) but avoids
setting the precedent of additional suffixed answers in

We'll likely want a helper to fetch all of the answers for a given
question for consumers showing or exporting data but that can be added
when the form runner/ export methods consume these helpers (alongside a
helper to show a "summary" of each entry in an add another which might
be configured by the form designer).