Switch to using FS_PLATFORM_ADMINS group
We now have a new Azure AD group specifically for platform admins on the new
Funding Service platform. Previously we had an FSD_ADMINS group, but this was
for more than just platform admins - it's also for form designers or any generic
  'funding service' staff/users.

We want to very tightly lock down access to the platform admin panel of the
integrated service, so this new group will represent access to that.

All other users will be set up within the service itself, using the service's
authorisation models (user/userrole).

After this is merged, anyone with the 'platform admin' permission in the system
will have it stripped away if they login and aren't a member of
FS_PLATFORM_ADMINS. This will be quite a few of our users; we can manually
migrate them using the platform admin user role page.