Automatically seed MHCLG organisation
We are now assocaiting grants with a 'grant owning' organisation. For now I'm
hardcoding a limit of a single grant-owning organisation within the whole
platform, and we'll start off with this being MHCLG. Knowing that only a single
org can be set up this way allows us to sidestep a number of design and
technical considerations, pushing those back to be dealt with when we _actually_
want to bring on another grant-owning org. Some considerations here are things
like:

- Deliver grant funding needing to understand/be designed around multiple orgs.

- Potential design considerations around users in 2+ grant-managing orgs (which
  might happen for our form designers)

We will automatically seed the MHCLG organisation during app startup. Generally,
expecting data to be present in the system can lead to a bit of snowflake
handling and making testing trickier. I've tried to limit this very strictly
here, and recommend strongly against continuing this principle. When we get to a
place where we have multiple grant-owning orgs, hopefully we can go back to just
treating all orgs as data and not having them be 'systems-level' data that is
expected/required by the system.