FSPT-850: helper to generate a single line summary for add another
When there are multiple questions this helper can be used to generate a
single line summary of all of their answers - as this is just stored in
presentation options for now it might contain references to old
questions so the helper is set up to default to returning a summary of
everything if there is either no configuration or no valid
configuraiton.

When the presentation options are being configured, all of the groups questions
will be shown and selected if there are no current presentation options
or if the existing presentation options aren't relevant. If relevant question
ids are stored in the presentation options only those will be selected.