FSPT-781: Add property to track referenced data qids
If a managed expression is using references to other questions in the
expression itself ie. in the case of referenced contextual data. This
adds a new property to managed expressions to keep track of these
references so we can create ComponentReference and ensure referential
integrity.

Deriving these from the expression safe_qid fields on the managed
expression means we don't need to track these in the session context,
so can remove these redundant properties.