FSPT-781: Create ComponentReference for managed expr. references
Now we're consistently tracking the question IDs of questions that
are being used as references in an expression statement, we need to
create a ComponentReference row for each of these used in an expression.

For each expression_referenced_question_ids, we create a new
ComponentReference. In some cases this might mean we have 3
ComponentReference rows for one expression - eg. for a 'Between'
Condition we could have one linking the target question and the depends
on question, one linking the target question and the minimum_expression
referenced question and one linking the target question and the
maximum_expression referenced question.