Add a lock timeout on migrations
We don't want this to hang forever - if it cannot acquire the locks it needs to
run the migrations in a very short order, we should fail so that we are quickly
notified. This either means some transactions have locked the DB up and that's a
problem we need to investigate, or we can just retry the migrations during a
less busy time (eg in some future where maybe we're running reporting that can
hold locks for a little while - hopefully never happens though).

Principle nicked from GOV.UK Notify:
https://github.com/alphagov/notifications-api/blob/7545e455993c4e9228718334e869435cd99e4f53/migrations/env.py#L104C47-L108C60