Ensure user_role.organisation_id set if grant_id set
Whenever we're giving people permissions related to a grant, we need to make
sure we contextualise that within a specific organisation. This will let us work
out if the user is a grant policy (deliver) member, or a grant recipient
(access) member.

This adds a layer to the existing permissions model for Deliver grant funding.
Previously:

1) Platform admin - role=ADMIN, org=none, grant=none. In deliver, can manage all
   grants by all orgs.

2) Grant team member - role=MEMBER, org=none, grant=x. In deliver, can view that
   specific grant.

After this change, we end up with a few more possible permutations:

1) Platform admin - role=ADMIN, org=none, grant=none. In deliver, can manage all
   grants by all orgs.

2) Deliver org admins - role=ADMIN, org=x, grant=none. In deliver, can manage
   all grants owned by that specific org.

3) Theoretical deliver org member - role=MEMBER, org=x, grant=none. In deliver,
   can view all grants owned by that specific org.

4) Theoretical grant team admin - role=ADMIN, org=x, grant=y. In deliver, can
   manage that specific grant.

5) Grant team member - role=MEMBER, org=x, grant=y. In deliver, can view that
   specific grant.

Adds a DB data migration to make sure any existing user_roles meet this
criteria, and then adds a check constraint to enforce it from here on.