BAU: Switch email used for e2e magic link test
We recently noticed that e2e test emails to the main service account
are going undelivered due to an issue with the email account. This
creates a lot of noise in notify and makes it difficult to know if
the email failed to send because of something to do with our own
app or due to the email address itself.

This swaps it out for a distribution list email which sends to no one
but is still a valid email.