FSPT-781: Allow classes to be applied to context-aware wrapper
The context-aware editor wrapper, initiliased dynamically with JS, is
what allows us to display the yellow highlight for interpolated
contextual references on Deliver Grant Funding.

However when adding a 'Remove data' button as part of the managed
expression data referencing flow, it became clear that we needed to be
able to pass the context-aware wrapper specific classes if we were also
applying them to the element it wrapped (eg. if you're changing
the width of an input field), otherwise the overlays didn't match up
which impacted the layout on the page.

This adds a new data attribute in which we can specify dispaly classes
that should also be applied to the dynamically-built wrapper.