Remove old admin actions; revoke permissions+invitations
At the moment we can only revoke invitations by deleting them entirely. It'd be
nice to retain the record that an invitation was created, but still make it
unusable. We can do this by moving the `expires_at_utc` date forward.

Also makes it possible to remove all permissions from a user. This is a
catch-all and we have more fine-grained ability to do this from the User Role
tab.