FSPT-781: Include referenced data in e2e expressions
Adds an example of referencing data in a numerical validation and date
condition to the e2e tests to ensure the data referencing flow and
functionality work as expected.

This involved a small refactor to how we set up validation/expressions
in our test questions and how they're added on the pages, but this
simplifies some functions and reduces some duplication overall and should
make it easy to add more in future if we need to, with consistent
behaviour.