FSPT-850: use summary list macro instead of HTML
If the first entry in the summary list has no actions it causes
subsequent rows to have the actions crammed into a small (20%) space -
this is likely because we're using it as K:actions rather than
K:V:actions. Only when the app-* class is applied add an override for
this behaviour.