Add platform admin pages to make a grant live
There are still a number of core deliver grant funding processes that we need to
represent in order to make reporting rounds live and available to grant
recipients. Given our limited time and resource at the moment, we're making a
pragmatic decision to build these out in 'platform admin' pages temporarily,
until we can come back to integrate them fully into Deliver grant funding with
proper care and consideration for sensible design and user journeys that eg
onboarding members could use.

We'll record all of these steps in a 'reporting lifecycle' tasklist, to make it
clear what needs to be done and the state of those tasks.