Refactor test client parameterisation
A lot of our tests use multiple clients in order to test access controls. As we
add new clients, we haven't done well at applying all of the different levels of
authorisation to our tests, so some of them test with just a subset.

This patch adds some new fixtures that auto-parameterise our tests, which means
if we add a new client/authorisation check we can update one place and all of
  our tests will automatically be updated with the new access control checks.