FSPT-850: fix add another conditional pages
The add another index wasn't being passed into the checks that validate
if a question page should be shown before its loaded.

We do this check because the conditions that make that page visble might
have changed since the user was routed to the page.

Add a test to make sure this doesn't regress for add another scenarios.