Prepare organisations table
We will be loading our list of organisations by syncing with Delta's. Initially
manually through a bulk upload, but we should start getting a regular sync (via
private s3 bucket) that we can use to update our data nightly.

This patch adds columns to the organisation table that we'll need and updates
the platform admin CRUD views.

Finally this ties in a bulk-upsert page through the reporting lifecycle so that
we can easily insert/update records based on an export from Delta