FPST_850: coverage for form data and tweak question page heading
Addresses a quirk in the question page heading by adding a rule to check
that nested groups inside an add another group correctly only show the
add another suffix on their heading and not in the caption too.

This part of the macro probably wants to be exhaustively unit tested!