FSPT-850: move question page heading and caption to helper
The template for the question page heading and caption was starting to
get convoluted and hard to reason about with all of the different
options between:
- being a single or same page group
- having guidance or not having guidance
- being part of an add another context or not

This lead to some less often configured scenarios incorrecly showing the
wrong heading/ caption which was hard to track down.

This commit lifts that logic out of the template which hopefully helps
the code be easier to follow and lets us assert what we'd expect to see
in each scenario to avoid regressions.