FSPT-850: view submission works for add anotother
This commit needs to decide if we'll go for one summary list component
that knows how to do this between check your answers and the form runner
or if we're OK with two.

I think its likely you'd want to "step in" to see the grouped answers as
an assessor/ reviewer so the pattern might be different in the future
anyway.

For now we're using a separate macro with a similar strategy, the
headings, keys and actions for the view submission page are all
different too.