Rename tasks to sections
After putting this through user research we have good evidence that the word
'task' is confusing for form designers and grant recipients. "Sections" is the
other terminology we have strong confidence in within the team; this is a
refactor to apply that across the codebase.

Note that we still refer to a "Tasklist" in places, as this is often in
reference to the GOV.UK tasklist component.

This refactor does not affect the underlying DB model, which is still called a
"Form". We may want to tackle this at some point to get true consistency.