FSPT-851: CSV export for add another
Quick first pass at making CSV work for add another, for now this just
chooses to repeat columns for add another contexts up to the maximum
number of times any submission has provided answers to that question.

This could be hard to parse and run into column length issues at extreme
but should be predictable behaviour for the time being.

It's possible that form designers/ grant teams might want this to work in
different ways depending on their data output needs, we might be able to
support:
- separate CSV files for each add another context with a reference
  linking them
- combined cell values for add another contexts