FSPT-908: add toggle add another to question group create flow
Lets the form designer select if a question group should be asked
multiple times while creating it.

The one complication here is that they shouldn't be able to create a
question group set to be add another inside an existin add another
question group - for now we'll just avoid asking the add another
question in that case.