Prevent users inserting references to same-page questions
When a user is creating a question within a same-page group, we were showing
them all of the questions in that same-page group in the 'reference data from a
previous question in this section' dropdown.

These references are invalid: you cannot reference data from a question that
shows on the same page, because that question hasn't been answered yet.

We should exclude these questions from the dropdown.