Remove Collection versioning
Collection versioning has been something we expected to bring in for a while,
but we didn't ever get it over the line. At this point the code is adding
complexity that provides no functional benefit, so this patch removes it. This
will simplify the implementation of the Reports entity, which otherwise might
need to understand versions on the backend despite nothing on the frontend
alluding to or using them.

As/when we get versioninig prioritised properly we can bring this all back in,
intentionally, with appropriate design considerations.