Set dates on collections
This patch allows a platform admin to set reporting and submission dates for a
monitoring report, with some simple validation that these are chronologically
ordered.

A collection must have these dates set in order to go live/be scheduled to go
live.

I've taken a few shortcuts here and put the dates directly on the Collection
model. Theoretically we may want to have some kind of separate `Report` DB model
to hold these changes, and we'd probably want to also move the `name` attribute
from the collection to that model as well. However, I'm choosing not to do that
for a few overlapping reasons:

- We're short on time and this is much simpler; adding that interstitial model
  would require a chunky refactor to do it properly.

- We don't have anything other than reports which use collections right now;
  when we have concrete reasons to support applications, prospectus info, grant
  setup information etc - some other kind of collection basically - we will be
  able to reason through a sensible model more thoroughly.