Set up grant recipient users in platform admin
Allow platform admins to set up grant recipient users.

We have set up a spreadsheet that we'll share with grant teams for them to
populate this information. One of the sheets has been structured to allow it to
be copied and pasted into this new page to set up all of the grant recipient
users in a single pass.

If a grant team has added some organisations we don't have a record of yet,
platform admins will need to create those first.

If any grant recipient users are added accidentally, there is a link to a page
where those users can be revoked again.