FSPT-851: handle submissions with different sizes
If a user has provided less answers than the maximum of any submission
we shouldn't attempt to get their answers + should use "not asked" to
indicate they never added that many.

The CSV generation method currently has a lot of responsibility and has
gotten nested and complex - we should split it up.