BAU Add moving grants to onboarding to reporting lifecycle
This commit adds the ability to mark a grant as "onboarding" in the
reporting lifecycle.

Its the first grant step and doesn't have any requirements.

This allows our team to work on "Active" grants which are always
prioritised at the top of their dashboard while allowing others in the
team and the wider org to setup grants that are either coming far in the
future or just to get familiar with the tools.

We've added no hard state transition requirements to move to this status
after "Draft" and before "Live" because it doesn't feel like it should
be a hard requirement and does no checks of its own.