Prevent report changes when non-draft
When a report (collection) has been "signed off and locked" / "scheduled to go
live", we want to make sure that no more changes to it can be made as that would
undermine any QA done.

This patch adds protection on all of the 'edit report' (or any of the report's
sub entities such as groups, components, expressions etc) to check the report's
status. On a non-draft status it will redirect back to an appropriate 'view
report' page, even for platform admins.