Fix some test factory set up
One test wasn't setting the correct grant context (this test passes because the
endpoint doesn't actually ever check the grant, but the next patch will make it
do that).

The other fix is on the question factory and makes it pull its form property
from a `parent` component, if that has been set. Some questions/groups were
being passed a parent but not a form, which then made them get linked to random
new forms. Again this causes problems with the next patch which adds more
thorough requirements on the question being associated with the correct parent
entities.