Move grant details tab to end
In user research we've often seen the 'Grand details' tab as a landing/home page
causing confusion. There isn't a strong drive to do anything on this page, and
it's often not hugely relevant to the person's task.

We move grant details to the end of the service navigation here, and make the
'list reports' page the default landing page for a grant now.