FSPT-960: Temporarily skip magic link e2e test
We'd agreed we would tackle the e2e tests for Access grant funding
at the end of this initial build to preserve pace. The previous commit
adds extra validation and checks to the magic link flow to ensure the
provided email belongs to a user who exists and is part of a grant
recipient organisation, which means the basic e2e test would fail (as no
user consistently exists in local/deployed environments).

This makes some basic changes to the test & page assertions to match
new content and route pattern, but skips the test until we decide
how we want to tackle auth for magic link in e2e tests.