FSPT-960: App index redirects to Access index
Deliver grant funding and Access grant funding have two distinct and
separate user groups, and most users will go directly to the relevant
service. However, there's still a chance a user will try to access the
root of the service, in which case we need to make a decision about
where we redirect them or whether we show options for going through to
Access or Deliver to the user.

In the short term, we decided to redirect to the Access grant funding
index. The reasons for this are:
- Magic link login (which is what the Access grant funding index will in
turn redirect to) is set up to deal with internal users who might
have accidentally made their way there, and will send them on to the SSO
login route
- We want to minimise friction for external grant recipient users
- These users are also less likely to be aware of the difference between
Access and Deliver grant funding, so we should default to the service
they need

This does have implications for internal (mainly Funding Service) users
who might have bookmarked funding.communities.gov.uk and will then be
redirected to Access rather than Deliver as expected.