Allow revoking all data providers
We aren't sure that grant teams will give us the canonical list of all data
providers in a single pass, but we are planning to rely on the spreadsheet being
the source of truth in the first instance.

To make it a little less likely that someone can be given permission, removed
from the spreadsheet, and then not removed from the system, we add a checkbox to
start by revoking all existing data providers when importing new ones.