BAU: raise warnings if a non-explicit GET request can change state
As we're leaning heavily on the SQLAlchemy ORM models throughout our
code we should try and avoid situations where properties are unknowingly
commited. This doesn't need to be left up to developers making changes
and reviewers, require endpoints that should explicitly change state
with `GET` to be added to an explicit list.