FSPT-958: fixup styles for a standalone org nav
When there is no grant context but the user has multiple orgs (most
likely only on the select grants page) the user will be presented with
only the org nav.

The custom org nav styles and functionality assume that you've got a
grant nav below (to include the "Change organisation" link on mobile and
to style the border bottom etc.)

If there is no grant context but we still need to show the org nav make
sure its styled using the standard service nav and the mobile menu
functinality will work as standard.