Add a UK postcode validator
Adds a naive managed validation rule to check that a single line of text answers
conforms to, roughly, the UK postcode format.

This will allow invalid UK postcodes through (ie ones that don't resolve to a
real location), because we only match on the pattern. We don't do any lookups to
validate it's a real postcode.

Product are happy with this scope limitation at the moment.