FSPT-1006: Update Access index routing
Now we have the 'Change organisation' view, the index should direct
users with more than one grant recipient org to this page instead of
defaulting to the list_grants for all Access Grant Funding users.

Once we merge the list reports page we can add an extra branch
for users who are only part of 1 grant recipient organisation and that
org is only a grant recipient for 1 grant.

This commit also adds a small change to the `list_grants` view to sort
the grants alphabetically by name.