FSPT-965: Stub Grant Recipient Team page
It's a sensible assumption that grant recipient users will want to know
which other users have access to provide data or certify reports
on a grant by grant basis.

This commit adds a stub 'grant team' page for Access Grant Funding
which will surface all users who have the certify or data_provider
permissions at org level and at the specific grant level, and show their
permissions in a table.

This page has not yet been through the full UCD process so might change
in the future but this gets a stub page in time for UR and means the
full Access grant funding navigation now works as expected.