FSPT-1012: navigate directly to grant if only 1 in AGF
Updates the AGF index route to point to a single grant if you only have
one, you should never then see the select a grant page until your
organisation is associated with multiple grants.

Note if you arrive anauthorised with a link to a specific resource
you'll still be forwarded to that.