BAU: if an environment is behind basic auth include invite creds
For any external users that are accessing non-prod environments (usually
for specific testing purposes) include the environments credentials so
that we don't need to send them separately.

This will mean they have a record of those basic auth credentials but
they would otherwise be trusted with them to partipate in using the
environment in the first place.