Add grant recipients table to grant details
We (funding service / platform team) currently set up all of a grant's
recipients in a behind-the-scenes action. This is important first-class
information in the platform and we should make it more visible, especially to
grant team's who send us a spreadsheet with the users that need setting up. At
the moment grant teams have no way of validating we've done it correctly, so
this adds a table of grant recipients and their users to the grant details tab.
This will allow us to offload some of the responsibility for validating the
correct recipients and users have been set up to the grant team.